discard old format
